pr comments changes, moved gas pricer into separate package and used shortened client interfaces
